embed/requirements.txt
cohere
pillow
pinecone
python-dotenv
tqdm

